requirements.txt
Flask~=3.0.3
marshmallow~=3.21.3
requests~=2.32.3
python-dotenv~=1.0.1
httpie~=2.6.0
awscli~=1.23.0
pip